+++ gtk/gtkdragicon.c
@@ -120,8 +120,8 @@ gtk_drag_icon_native_get_renderer (GtkNative *native)
 
 static void
 gtk_drag_icon_native_get_surface_transform (GtkNative *native,
-                                            int       *x,
-                                            int       *y)
+                                            double    *x,
+                                            double    *y)
 {
   GtkCssBoxes css_boxes;
   const graphene_rect_t *margin_rect;
